requirements.txt
python-dotenv==1.0.0
pyupbit==0.2.33
pandas==2.0.3
numpy==1.24.3
ta==0.10.2
requests==2.31.0
pydantic==2.4.2
schedule==1.2.0
pytz==2023.3
telebot==0.0.5
mplfinance==0.12.10b0
matplotlib==3.7.2
torch==2.1.0
xgboost==2.0.0
scikit-learn==1.3.0
optuna==3.3.0
joblib==1.3.2